move the check to the right spot
